NAS-117581: Fix Launch Docker Image button bugs
