+++ src/app/pages/applications/catalog/catalog.component.ts
@@ -1,39 +1,28 @@
 import {
-  Component, OnInit, Output, EventEmitter, OnDestroy,
+  Component, OnInit, Output, EventEmitter, OnDestroy, Input,
 } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { Router } from '@angular/router';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { TranslateService } from '@ngx-translate/core';
 import { Subscription } from 'rxjs';
 import {
   chartsTrain, ixChartApp, officialCatalog, appImagePlaceholder,
 } from 'app/constants/catalog.constants';
 import { CommonUtils } from 'app/core/classes/common-utils';
-import { CoreService } from 'app/core/services/core-service/core.service';
 import { JobState } from 'app/enums/job-state.enum';
 import helptext from 'app/helptext/apps/apps';
-import { ApplicationUserEventName } from 'app/interfaces/application.interface';
+import { ApplicationUserEvent, ApplicationUserEventName } from 'app/interfaces/application.interface';
 import { CatalogApp } from 'app/interfaces/catalog.interface';
 import { CoreEvent } from 'app/interfaces/events';
-import { Job } from 'app/interfaces/job.interface';
-import { KubernetesConfig } from 'app/interfaces/kubernetes-config.interface';
 import { Option } from 'app/interfaces/option.interface';
-import { ApplicationTab } from 'app/pages/applications/application-tab.enum';
 import { ApplicationToolbarControl } from 'app/pages/applications/application-toolbar-control.enum';
-import { KubernetesSettingsComponent } from 'app/pages/applications/kubernetes-settings/kubernetes-settings.component';
-import { DialogFormConfiguration } from 'app/pages/common/entity/entity-dialog/dialog-form-configuration.interface';
-import { EntityDialogComponent } from 'app/pages/common/entity/entity-dialog/entity-dialog.component';
+import { ChartWizardComponent } from 'app/pages/applications/forms/chart-wizard.component';
 import { EmptyConfig, EmptyType } from 'app/pages/common/entity/entity-empty/entity-empty.component';
 import { EntityJobComponent } from 'app/pages/common/entity/entity-job/entity-job.component';
-import { EntityUtils } from 'app/pages/common/entity/utils';
 import { AppLoaderService } from 'app/services/app-loader/app-loader.service';
 import { DialogService, WebSocketService } from 'app/services/index';
-import { IxSlideInService } from 'app/services/ix-slide-in.service';
 import { ModalService } from 'app/services/modal.service';
 import { ApplicationsService } from '../applications.service';
 import { CatalogSummaryDialogComponent } from '../dialogs/catalog-summary/catalog-summary-dialog.component';
-import { ChartWizardComponent } from '../forms/chart-wizard.component';
 
 interface CatalogSyncJob {
   id: number;
@@ -48,7 +37,8 @@ interface CatalogSyncJob {
   styleUrls: ['../applications.component.scss', 'catalog.component.scss'],
 })
 export class CatalogComponent implements OnInit, OnDestroy {
-  @Output() updateTab = new EventEmitter();
+  @Input() selectedPool: string;
+  @Output() updateTab: EventEmitter<ApplicationUserEvent> = new EventEmitter();
 
   catalogApps: CatalogApp[] = [];
   catalogNames: string[] = [];
@@ -56,8 +46,6 @@ export class CatalogComponent implements OnInit, OnDestroy {
   filteredCatalogApps: CatalogApp[] = [];
   filterString = '';
   catalogSyncJobs: CatalogSyncJob[] = [];
-  selectedPool = '';
-  private poolList: Option[] = [];
 
   jobsSubscription: Subscription;
 
@@ -73,45 +61,19 @@ export class CatalogComponent implements OnInit, OnDestroy {
 
   readonly officialCatalog = officialCatalog;
 
-  choosePool: DialogFormConfiguration = {
-    title: helptext.choosePool.title,
-    fieldConfig: [
-      {
-        type: 'select',
-        name: 'pools',
-        placeholder: helptext.choosePool.placeholder,
-        required: true,
-        options: this.poolList,
-      },
-      {
-        type: 'checkbox',
-        name: 'migrateApplications',
-        placeholder: helptext.choosePool.migrateApplications,
-      },
-    ],
-    method_ws: 'kubernetes.update',
-    saveButtonText: helptext.choosePool.action,
-    customSubmit: (entityForm) => this.doPoolSelect(entityForm),
-  };
-
   constructor(
     private dialogService: DialogService,
     private appLoaderService: AppLoaderService,
     private mdDialog: MatDialog,
-    private translate: TranslateService,
     private ws: WebSocketService,
-    private router: Router,
-    private core: CoreService,
     private modalService: ModalService,
     private appService: ApplicationsService,
-    private slideInService: IxSlideInService,
   ) {
     this.utils = new CommonUtils();
   }
 
   ngOnInit(): void {
     this.loadCatalogs();
-    this.checkForConfiguredPool();
 
     this.jobsSubscription = this.ws.subscribe('core.get_jobs').pipe(untilDestroyed(this)).subscribe((event) => {
       const catalogSyncJob = this.catalogSyncJobs.find((job) => job.id == event.fields.id);
@@ -143,6 +105,7 @@ export class CatalogComponent implements OnInit, OnDestroy {
     this.appService.getAllCatalogItems().pipe(untilDestroyed(this)).subscribe((catalogs) => {
       this.noAvailableCatalog = true;
       this.catalogApps = [];
+      this.catalogNames = [];
       catalogs.forEach((catalog) => {
         if (!catalog.cached) {
           if (catalog.caching_job) {
@@ -210,21 +173,7 @@ export class CatalogComponent implements OnInit, OnDestroy {
   }
 
   onToolbarAction(evt: CoreEvent): void {
-    if (evt.data.event_control === ApplicationToolbarControl.Settings && evt.data.settings) {
-      switch (evt.data.settings.value) {
-        case 'select_pool':
-          this.selectPool();
-          return;
-        case 'advanced_settings':
-          this.slideInService.open(KubernetesSettingsComponent);
-          break;
-        case 'unset_pool':
-          this.doUnsetPool();
-          break;
-      }
-    } else if (evt.data.event_control === ApplicationToolbarControl.Launch) {
-      this.doInstall(ixChartApp);
-    } else if (evt.data.event_control === ApplicationToolbarControl.Filter) {
+    if (evt.data.event_control === ApplicationToolbarControl.Filter) {
       this.filterString = evt.data.filter;
       this.filterApps();
     } else if (evt.data.event_control === ApplicationToolbarControl.RefreshAll) {
@@ -237,123 +186,7 @@ export class CatalogComponent implements OnInit, OnDestroy {
   }
 
   refreshToolbarMenus(): void {
-    this.updateTab.emit({
-      name: ApplicationUserEventName.CatalogToolbarChanged,
-      value: Boolean(this.selectedPool),
-      catalogNames: this.catalogNames,
-    });
-  }
-
-  checkForConfiguredPool(): void {
-    this.appService.getKubernetesConfig().pipe(untilDestroyed(this)).subscribe((config) => {
-      if (!config.pool) {
-        this.selectPool();
-        this.updateTab.emit({ name: ApplicationUserEventName.SwitchTab, value: ApplicationTab.AvailableApps });
-      } else {
-        this.selectedPool = config.pool;
-      }
-      this.refreshToolbarMenus();
-    });
-  }
-
-  selectPool(): void {
-    this.appService.getPoolList().pipe(untilDestroyed(this)).subscribe((pools) => {
-      if (pools.length === 0) {
-        this.dialogService.confirm({
-          title: helptext.noPool.title,
-          message: helptext.noPool.message,
-          hideCheckBox: true,
-          buttonMsg: helptext.noPool.action,
-        }).pipe(untilDestroyed(this)).subscribe((confirmed) => {
-          if (!confirmed) {
-            return;
-          }
-          this.router.navigate(['storage', 'manager']);
-        });
-      } else {
-        this.poolList.length = 0;
-        pools.forEach((pool) => {
-          this.poolList.push({ label: pool.name, value: pool.name });
-        });
-
-        const migrateField = this.choosePool.fieldConfig.find((config) => config.name === 'migrateApplications');
-        if (this.selectedPool) {
-          this.choosePool.fieldConfig[0].value = this.selectedPool;
-          migrateField.isHidden = false;
-        } else {
-          delete this.choosePool.fieldConfig[0].value;
-          migrateField.isHidden = true;
-        }
-
-        this.dialogService.dialogForm(this.choosePool, true);
-      }
-    });
-  }
-
-  doUnsetPool(): void {
-    this.dialogService.confirm({
-      title: helptext.choosePool.unsetPool.confirm.title,
-      message: helptext.choosePool.unsetPool.confirm.message,
-      hideCheckBox: true,
-      buttonMsg: helptext.choosePool.unsetPool.confirm.button,
-    }).pipe(untilDestroyed(this)).subscribe((confirmed) => {
-      if (!confirmed) {
-        return;
-      }
-
-      const dialogRef = this.mdDialog.open(EntityJobComponent, {
-        data: {
-          title: helptext.choosePool.jobTitle,
-        },
-        disableClose: true,
-      });
-      dialogRef.componentInstance.setCall('kubernetes.update', [{ pool: null }]);
-      dialogRef.componentInstance.submit();
-      dialogRef.componentInstance.success.pipe(untilDestroyed(this)).subscribe(() => {
-        this.dialogService.closeAllDialogs();
-        this.selectedPool = null;
-        this.refreshToolbarMenus();
-        this.dialogService.info(
-          helptext.choosePool.success,
-          this.translate.instant(helptext.choosePool.unsetPool.label),
-          '500px',
-          'info',
-          true,
-        );
-      });
-
-      dialogRef.componentInstance.failure.pipe(untilDestroyed(this)).subscribe((err) => {
-        new EntityUtils().handleWsError(this, err, this.dialogService);
-      });
-    });
-  }
-
-  doPoolSelect(entityDialog: EntityDialogComponent): void {
-    const pool = entityDialog.formGroup.controls['pools'].value;
-    const migrateApplications = entityDialog.formGroup.controls['migrateApplications'].value;
-    this.dialogService.closeAllDialogs();
-    const dialogRef = this.mdDialog.open(EntityJobComponent, {
-      data: {
-        title: helptext.choosePool.jobTitle,
-      },
-    });
-    dialogRef.componentInstance.setCall('kubernetes.update', [{
-      pool,
-      migrate_applications: migrateApplications,
-    }]);
-    dialogRef.componentInstance.submit();
-    dialogRef.componentInstance.success.pipe(untilDestroyed(this)).subscribe((res: Job<KubernetesConfig>) => {
-      this.selectedPool = pool;
-      this.refreshToolbarMenus();
-      this.dialogService.closeAllDialogs();
-      this.dialogService.info(
-        helptext.choosePool.success,
-        this.translate.instant(helptext.choosePool.message) + res.result.pool,
-        '500px',
-        'info',
-        true,
-      );
-    });
+    this.updateTab.emit({ name: ApplicationUserEventName.CatalogToolbarChanged, catalogNames: this.catalogNames });
   }
 
   doInstall(name: string, catalog = officialCatalog, train = chartsTrain): void {
